Move MongoDB specific methods to dao module
